Configure testing

+++ server/notesFeature/controller.js
@@ -1,12 +1,11 @@
 const Notes = require('./model')
+const isNormalInteger = require('./util')
 
 function createNote(req, res) {
   Notes.create({ ...req.body }, (err, note) => {
     if (err || !note) {
-      console.log({ err })
       return res.status(400).end(err)
     } else {
-      console.log({ note })
       return res.status(200).send(note)
     }
   })
@@ -14,21 +13,26 @@ function createNote(req, res) {
 
 function getNoteFromUid(req, res) {
   const { uid } = req.params
-  Notes.findOne({ uid }, (err, note) => {
-    if (err || !note) {
-      console.log({ err })
-      return res.status(400).end(err)
-    } else {
-      console.log({ note })
-      return res.status(200).send(note)
-    }
-  })
+  if (isNormalInteger(uid)) {
+    Notes.findOne({ uid }, (err, note) => {
+      if (err || !note) {
+        return res.status(400).end(err)
+      } else {
+        return res.status(200).send(note)
+      }
+    })
+  } else {
+    res.status(400).end(
+      `The value ${uid} is not a recognized format. 
+      getNoteFromUid requires either '0' or a positive integer.
+    `
+    )
+  }
 }
 
 function getAllNotes(req, res) {
   Notes.find({}, (err, notes) => {
     if (err || !notes) {
-      console.log({ err })
       return res.status(400).end(err)
     } else {
       console.log({ notes })
@@ -39,37 +43,45 @@ function getAllNotes(req, res) {
 
 function updateNoteFromUid(req, res) {
   const { uid } = req.params
-  const { property, replace } = req.body
 
-  Notes.findOneAndUpdate(
-    { uid },
-    {
-      [property]: replace
-    },
-    { new: true },
-    (err, note) => {
-      if (err || !note) {
-        console.log({ err })
-        return res.status(400).end(err)
-      } else {
-        console.log({ note })
-        return res.status(200).send(note)
+  if (isNormalInteger(uid)) {
+    const { property, replace } = req.body
+    Notes.findOneAndUpdate(
+      { uid },
+      {
+        [property]: replace
+      },
+      { new: true },
+      (err, note) => {
+        if (err || !note) {
+          return res.status(400).end(err)
+        } else {
+          return res.status(200).send(note)
+        }
       }
-    }
-  )
+    )
+  } else {
+  }
 }
 
 function deleteNoteFromUid(req, res) {
   const { uid } = req.params
-  Notes.findOneAndRemove({ uid }, (err, note) => {
-    if (err || !note) {
-      console.log({ err })
-      return res.status(400).end(err)
-    } else {
-      console.log({ note })
-      return res.status(200).send(note)
-    }
-  })
+
+  if (isNormalInteger(uid)) {
+    Notes.findOneAndRemove({ uid }, (err, note) => {
+      if (err || !note) {
+        return res.status(400).end(err)
+      } else {
+        return res.status(200).send(note)
+      }
+    })
+  } else {
+    res.status(400).end(
+      `The value ${uid} is not a recognized format. 
+      getNoteFromUid requires either '0' or a positive integer.
+    `
+    )
+  }
 }
 
 module.exports = {
